# shop-api/docs/swagger.yaml
basePath: /
definitions:
  models.CartAddRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  models.CartUpdateRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  models.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.UserAddress:
    properties:
      address:
        type: string
      city:
        type: string
      consignee:
        type: string
      created_at:
        type: string
      district:
        type: string
      id:
        type: integer
      is_default:
        description: 0=非默认,1=默认
        type: integer
      mobile:
        type: string
      province:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 爱影视频APP商城系统和第三方站点管理API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 爱影视频商城API
  version: "1.0"
paths:
  /api/v1/shop/addresses:
    get:
      consumes:
      - application/json
      description: 获取当前用户的地址列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取地址列表
      tags:
      - 地址
    post:
      consumes:
      - application/json
      description: 创建新地址
      parameters:
      - description: 地址信息
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.UserAddress'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建地址
      tags:
      - 地址
  /api/v1/shop/addresses/{id}:
    delete:
      consumes:
      - application/json
      description: 删除地址
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除地址
      tags:
      - 地址
    get:
      consumes:
      - application/json
      description: 根据ID获取地址详情
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取地址详情
      tags:
      - 地址
    put:
      consumes:
      - application/json
      description: 更新地址信息
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      - description: 地址信息
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.UserAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新地址
      tags:
      - 地址
  /api/v1/shop/cart:
    get:
      consumes:
      - application/json
      description: 获取当前用户的购物车
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取购物车
      tags:
      - 购物车
    post:
      consumes:
      - application/json
      description: 添加商品到购物车
      parameters:
      - description: 商品信息
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/models.CartAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 添加商品到购物车
      tags:
      - 购物车
  /api/v1/shop/cart/{id}:
    delete:
      consumes:
      - application/json
      description: 从购物车中移除商品
      parameters:
      - description: 购物车项ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 从购物车移除商品
      tags:
      - 购物车
    put:
      consumes:
      - application/json
      description: 更新购物车中商品的数量
      parameters:
      - description: 购物车项ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/models.CartUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新购物车商品数量
      tags:
      - 购物车
  /api/v1/shop/cart/clear:
    delete:
      consumes:
      - application/json
      description: 清空当前用户的购物车
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 清空购物车
      tags:
      - 购物车
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"